+++ include/input.hpp
@@ -1,0 +1,9 @@
+#ifndef INCLUDE_GUARD_INPUT_HPP
+#define INCLUDE_GUARD_INPUT_HPP
+
+#include <string>
+#include "./boardpoint.hpp"
+
+BoardPoint inputTranslator(const std::string input);
+
+#endif  // INCLUDE_GUARD_INPUT_HPP
