+++ include/output.hpp
@@ -1,0 +1,11 @@
+#ifndef INCLUDE_GUARD_OUTPUT_HPP
+#define INCLUDE_GUARD_OUTPUT_HPP
+
+#include "./fieldinfo.hpp"
+
+void clearScreen();
+void printLine();
+void printNumbers();
+void printGameBoard(const BoardType& board);
+
+#endif  // INCLUDE_GUARD_OUTPUT_HPP
